Change notation
Earlier I was using x \in R^n and u \in R^m. This was for consistency
with the MDP part of the notes where n and m are the size of the state
and the action space. However, this notation means that we need to come
up with another variable for observations.

Moving to the "standard" notation that we use in papers to avoid having
too many letters just to denote dimension of variables